+++ app/restaurant/[slug]/components/ReservationCard.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+"use client"
 
 export default function ReservationCard() {
   return (
