+++ app/restaurant/[slug]/page.tsx
@@ -1,4 +1,3 @@
-import NavBar from "../../components/NavBar";
 import Header from "./components/Header";
 import RestaurantNavBar from "./components/RestaurantNavBar";
 import Title from "./components/Title";
@@ -10,29 +9,21 @@ import ReservationCard from "./components/ReservationCard";
 
 export default function RestaurantDetails() {
   return (
-    <main className="bg-gray-100 min-h-screen w-screen">
-      <main className="max-w-screen-2xl m-auto bg-white">
-        <NavBar />
-        <Header />
-        <div className="flex m-auto w-2/3 justify-between items-start 0 -mt-11">
-          <div className="bg-white w-[70%] rounded p-3 shadow">
-            <RestaurantNavBar />
-            <Title />
-            <Rating />
-           <Description />
-            <Images />
-            {/* REVIEWS */}
-            <Reviews />
-            {/* REVIEWS */}
-          </div>
-          <div className="w-[27%] relative text-reg">
-            <ReservationCard />
-          </div>
+    <>
+      <Header />
+      <div className="flex m-auto w-2/3 justify-between items-start 0 -mt-11">
+        <div className="bg-white w-[70%] rounded p-3 shadow">
+          <RestaurantNavBar />
+          <Title />
+          <Rating />
+          <Description />
+          <Images />
+          <Reviews />
         </div>
-        {/* DESCRIPTION PORTION */} {/* RESERVATION CARD PORTION */}{" "}
-        {/* RESERVATION
-          CARD PORTION */}
-      </main>
-    </main>
+        <div className="w-[27%] relative text-reg">
+          <ReservationCard />
+        </div>
+      </div>
+    </>
   );
 }
